Fixed issue with artifact client Switched to typescript
