+++ package.json
@@ -6,7 +6,7 @@
   "author": "NinjaManatee",
   "license": "MIT",
   "scripts": {
-    "build": "ncc build ./src/index.js --license licenses.txt",
+    "build": "ncc build ./src/index.ts --license licenses.txt",
     "postinstall": "npm run build"
   },
   "dependencies": {
@@ -15,6 +15,8 @@
     "screenshot-desktop": "^1.15.1"
   },
   "devDependencies": {
-    "@vercel/ncc": "^0.38.3"
+    "@types/screenshot-desktop": "^1.12.3",
+    "@vercel/ncc": "^0.38.3",
+    "typescript": "^5.7.3"
   }
 }
