+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+        "allowSyntheticDefaultImports": true,
+        "moduleResolution": "node",
+        "target": "ES2024",
+    },
+    "files": ["./src/index.ts"]
+}
